==> notebooks/atlas/06_AtlasAge_routines.ipynb <==
# Debugging autoreload

%load_ext autoreload
%autoreload 2

# Load packages

from pytorch_tabular.utils import load_covertype_dataset
from rich.pretty import pprint
from sklearn.model_selection import BaseCrossValidator, ParameterGrid, ParameterSampler
from pandas.api.types import is_numeric_dtype
import torch
import pickle
import shutil
import shap
from sklearn.model_selection import RepeatedStratifiedKFold
from glob import glob
import ast
import matplotlib.pyplot as plt
import seaborn as sns
import copy
from sklearn.model_selection import train_test_split
import numpy as np
from pytorch_tabular.utils import make_mixed_dataset, print_metrics
from pytorch_tabular import available_models
from pytorch_tabular import TabularModel
from pytorch_tabular.models import CategoryEmbeddingModelConfig, GANDALFConfig, TabNetModelConfig, FTTransformerConfig, DANetConfig
from pytorch_tabular.config import DataConfig, OptimizerConfig, TrainerConfig
from pytorch_tabular.models.common.heads import LinearHeadConfig
from pytorch_tabular.tabular_model_tuner import TabularModelTuner
from sklearn.metrics import mean_absolute_error
from pytorch_tabular import MODEL_SWEEP_PRESETS
import pandas as pd
from pytorch_tabular import model_sweep
from src.pt.model_sweep import model_sweep_custom
import warnings
from src.utils.configs import read_parse_config
from src.utils.hash import dict_hash
from src.pt.hyper_opt import train_hyper_opt
import pathlib
from tqdm import tqdm
import distinctipy
import matplotlib.patheffects as pe
import matplotlib.colors as mcolors
from statannotations.Annotator import Annotator
from scipy.stats import mannwhitneyu
from regression_bias_corrector import LinearBiasCorrector
import optuna
from sklearn.preprocessing import LabelEncoder
from plottable import ColumnDefinition, Table
from plottable.plots import bar
from plottable.cmap import normed_cmap, centered_cmap
import matplotlib.lines as mlines
import plotly.graph_objects as go
from plotly.subplots import make_subplots
import scipy.stats
from functools import reduce
from sklearn.impute import KNNImputer
from scipy import stats
import upsetplot
from docx import Document
from docx.shared import Inches, Cm, Mm, Pt
from docx.enum.text import WD_ALIGN_PARAGRAPH, WD_BREAK
from docx.enum.section import WD_ORIENT
from docx.enum.style import WD_STYLE_TYPE
from docx.shared import RGBColor


def make_rgb_transparent(rgb, bg_rgb, alpha):
    return [alpha * c1 + (1 - alpha) * c2 for (c1, c2) in zip(rgb, bg_rgb)]


def delete_paragraph(paragraph):
    p = paragraph._element
    p.getparent().remove(p)
    p._p = p._element = None
    

def markdown_to_docx(markdown_text, doc):
    
    # Регулярные выражения для обработки элементов
    header_re = re.compile(r'^(#+)\s+(.*)')
    list_re = re.compile(r'^(\s*)- (.*)')
    bold_re = re.compile(r'(\*\*|__)(.*?)\1')
    italic_re = re.compile(r'(\*|_)(.*?)\1')
    
    # Обработка каждой строки
    for line in markdown_text.split('\n'):
        stripped_line = line.strip()
        
        # Пропуск пустых строк и разделителей
        if not stripped_line or re.match(r'^-{3,}$', stripped_line):
            continue
        
        # Обработка заголовков
        header_match = header_re.match(line)
        if header_match:
            level = len(header_match.group(1))
            text = header_match.group(2).strip()
            doc.add_heading(text, level=min(level, 6))
            continue
        
        # Обработка списков
        if re.match(r'^[\s]*[-*+] ', line):
            indent = len(line) - len(line.lstrip())
            p = doc.add_paragraph(style='ListBullet')
            p.paragraph_format.left_indent = Pt(25 + 10 * (indent//2))
            line_content = line.lstrip()[2:]
            add_formatted_text(p, line_content, bold_re, italic_re)
            continue
        
        # Обработка обычного текста
        p = doc.add_paragraph()
        add_formatted_text(p, line.strip(), bold_re, italic_re)
    
    return doc

def add_formatted_text(paragraph, text, bold_re, italic_re):
    # Добавление текста с форматированием
    fragments = []
    last_pos = 0
    
    # Разделение текста на фрагменты с форматированием
    for match in bold_re.finditer(text):
        start, end = match.start(), match.end()
        if last_pos < start:
            fragments.append(('normal', text[last_pos:start]))
        fragments.append(('bold', match.group(2)))
        last_pos = end
    
    if last_pos < len(text):
        fragments.append(('normal', text[last_pos:]))
    
    # Обработка курсива внутри оставшихся фрагментов
    final_fragments = []
    for frag_type, content in fragments:
        if frag_type == 'bold':
            final_fragments.append(('bold', content))
            continue
            
        sub_last = 0
        for match in italic_re.finditer(content):
            start, end = match.start(), match.end()
            if sub_last < start:
                final_fragments.append(('normal', content[sub_last:start]))
            final_fragments.append(('italic', match.group(2)))
            sub_last = end
        
        if sub_last < len(content):
            final_fragments.append(('normal', content[sub_last:]))
    
    # Добавление фрагментов в параграф
    for frag_type, content in final_fragments:
        run = paragraph.add_run(content)
        if frag_type == 'bold':
            run.bold = True
        elif frag_type == 'italic':
            run.italic = True


# Load data and models for subsets of features

path = f"E:/YandexDisk/Work/bbd/atlas"

feat_trgt = 'Возраст'

data_suffix = '_v3'
data = pd.read_excel(f"{path}/data{data_suffix}.xlsx", index_col=0)

components = {
    'InBody': {
        'name': 'Биоимпеданс (InBody)',
        'path': f"{path}/subset_InBody-mRMR_no-sex",
        'path_model': f"{path}/subset_InBody-mRMR_no-sex/models/DANet/2/432", 
        'color': 'dodgerblue',
        'bkg_count': 50,
        'likelihood': 1.0
        
    },
    'CompleteBloodCount': {
        'name': 'Общий анализ крови',
        'path': f"{path}/subset_CBC_no-sex",
        'path_model': f"{path}/subset_CBC_no-sex/models/DANet/2/67", 
        'color': 'crimson',
        'bkg_count': 300,
        'likelihood': 1.0
    },
    'BloodBiochemical': {
        'name': 'Биохимия крови',
        'path': f"{path}/subset_BloodBiochemical_no-sex",
        'path_model': f"{path}/subset_BloodBiochemical_no-sex/models/DANet/1/630", 
        'color': 'cyan',
        'bkg_count': 150,
        'likelihood': 1.0
    },
    "LipidProfile": {
        "name": "Липидный профиль",
        "path": f"{path}/subset_LipidProfile_no-sex",
        "path_model": f"{path}/subset_LipidProfile_no-sex/models/DANet/1/798",
        "color": "gold",
        'bkg_count': 100,
        'likelihood': 0.75
    },
    "CoagulationTest": {
        "name": "Коагулограмма",
        "path": f"{path}/subset_CoagulationTest_no-sex",
        "path_model": f"{path}/subset_CoagulationTest_no-sex/models/DANet/1/868",
        "color": "olive",
        'bkg_count': 100,
        'likelihood': 0.75
    },
    "HormoneProfile": {
        "name": "Гормональный профиль",
        "path": f"{path}/subset_HormoneProfile_no-sex",
        "path_model": f"{path}/subset_HormoneProfile_no-sex/models/DANet/1/425",
        "color": "chocolate",
        'bkg_count': 50,
        'likelihood': 0.6
    },
    "ProstateSpecificAntigenTest": {
        "name": "Простатический специфический антиген",
        "path": f"{path}/subset_ProstateSpecificAntigenTest_no-sex",
        "path_model": f"{path}/subset_ProstateSpecificAntigenTest_no-sex/models/DANet/1/529",
        "color": "lawngreen",
        'bkg_count': 50,
        'likelihood': 0.6
    },
    "RheumatologyScreening": {
        "name": "Ревматологический тест",
        "path": f"{path}/subset_RheumatologyScreening_no-sex",
        "path_model": f"{path}/subset_RheumatologyScreening_no-sex/models/DANet/1/48",
        "color": "gray",
        'bkg_count': 50,
        'likelihood': 0.5
    },
    # "BloodPressure": {
    #     "name": "Кровяное Давление",
    #     "path": f"{path}/subset_BloodPressure_no-sex",
    #     "path_model": f"{path}/subset_BloodPressure_no-sex/models/DANet/1/75",
    #     "color": "orchid",
    #     'bkg_count': 300,
    #     'likelihood': 0.6
    # },
}

feats_all = []
for comp in components:
    components[comp]['data'] = pd.read_excel(f"{components[comp]['path']}/data.xlsx", index_col=0)
    components[comp]['feats'] = pd.read_excel(f"{components[comp]['path']}/feats.xlsx", index_col=0)
    components[comp]['results'] = pd.read_excel(f"{components[comp]['path_model']}/df.xlsx", index_col=0)
    components[comp]['metrics'] = pd.read_excel(f"{components[comp]['path_model']}/metrics.xlsx", index_col=0)
    components[comp]['shap'] = pd.read_excel(f"{components[comp]['path_model']}/explanation.xlsx", index_col=0)
    components[comp]['model'] = TabularModel.load_model(f"{components[comp]['path_model']}")
    components[comp]['corrector'] = LinearBiasCorrector()
    comp_results = components[comp]['results']
    components[comp]['corrector'].fit(comp_results.loc[comp_results['Group'] == 'Train', feat_trgt].values, comp_results.loc[comp_results['Group'] == 'Train', 'Prediction'].values)
    res_cols = ['Group', 'Prediction', 'Error', 'Prediction Unbiased', 'Error Unbiased']
    components[comp]['data'].loc[components[comp]['data'].index, res_cols] = comp_results.loc[components[comp]['data'].index, res_cols]
    components[comp]['data_shap'] = components[comp]['data'].copy()
    
    feats = components[comp]['feats'].index.values
    feats = feats[feats != feat_trgt]
    feats_all += list(feats)
    feats_all += [f"Предсказание {components[comp]['name']}", f"Возрастная Акселерация {components[comp]['name']}"]
    
    components[comp]['feats_corr'] = pd.DataFrame(index=feats, columns=['Correlation'])
    for f in feats:
        components[comp]['feats_corr'].at[f, 'Correlation'], _ = stats.pearsonr(components[comp]['data'].loc[:, f].values, components[comp]['data'].loc[:, feat_trgt].values)

for comp in components:
    print(f"{comp}: {components[comp]['data'].shape[0]}")

## Raw data

data = pd.read_excel(f"{path}/AtlasAge/data_5.xlsx", index_col=0)

# Возраст
if not is_numeric_dtype(data['Возраст']):
    print("Processing Age")
    age_parts = data['Возраст'].astype(str).str.extract(r"^(\d+)\D*(\d*)\D*$")
    age_parts.columns = ['Years', 'Month']
    age_parts['Month'] = age_parts['Month'].replace({'': 0})
    age_parts['Возраст'] = pd.to_numeric(age_parts['Years']) + pd.to_numeric(age_parts['Month']) / 12.0
    data.loc[data.index, 'Возраст'] = age_parts.loc[data.index, 'Возраст']
    data['Возраст'] = pd.to_numeric(data['Возраст'])
    
# Давление
if not {'Cистолическое АД', 'Диастолическое АД'}.issubset(data.columns) and 'АД' in data.columns:
    print("Processing Blood Pressure")
    blood_pressure = data['АД'].astype(str).str.extract(r"^(\d+)/(\d+)$")
    blood_pressure.columns = ['Cистолическое АД', 'Диастолическое АД']
    data.loc[data.index, ['Cистолическое АД', 'Диастолическое АД']] = blood_pressure.loc[data.index, ['Cистолическое АД', 'Диастолическое АД']]
    data['Cистолическое АД'] = pd.to_numeric(data['Cистолическое АД'])
    data['Диастолическое АД'] = pd.to_numeric(data['Диастолическое АД'])
    
# % свободного ПСА
if '% свободного ПСА, %' in data.columns:
    data.rename(columns={'% свободного ПСА, %': '% свободного ПСА'}, inplace=True)

# Много знаков %
data.columns = data.columns.str.replace(" %, %", ", %", regex=True)
data.columns = data.columns.str.replace(",,", ",", regex=True)

# Двойные пробелы
data.columns = data.columns.str.replace("  ", " ", regex=True)

# Двойные пробелы
data.columns = data.columns.str.replace("  ", " ", regex=True)

# Запятая c пробелом в конце названия столбца
data.columns = [x[:-2] if x[-2:] == ', ' else x for x in data.columns]

# Пробел перед запятой
data.columns = data.columns.str.replace(" ,", ",", regex=True)

# Replace nan-characters with nan
data.replace({'.': np.nan, }, inplace=True)
data.replace({'-': np.nan, }, inplace=True)
data.replace({r"^.*Необходимо повторить исследование.*$": np.nan}, regex=True, inplace=True)
data.replace({r"^.*расчет невозможен.*$": np.nan}, regex=True, inplace=True)
data.replace({r"^.*нет результата.*$": np.nan}, regex=True, inplace=True)
data.replace({r"^.*без особенностей.*$": np.nan}, regex=True, inplace=True)
data.replace({r'^([<>].*)$': np.nan}, regex=True, inplace=True)

if 'Тироксин свободный (Т4 свободный), пмоль/л' in data.columns:
    data['Тироксин свободный (Т4 свободный), пмоль/л'] = pd.to_numeric(data['Тироксин свободный (Т4 свободный), пмоль/л'], errors='coerce')

# Уникальность столбцов
cols_dupl = data.columns.values[data.columns.duplicated()]
if len(cols_dupl) > 0:
    print(cols_dupl)
    
# Проверка признаков в моделях и в данных
for comp in components:
    comp_only_feats = list(set(components[comp]['feats'].index.values) - set(data.columns.values))
    print(f"{comp}: {comp_only_feats}")

{'Cистолическое АД', 'Диастолическое АД', 'АД'}.issubset(data.columns)

# Prediction for samples

nan_part = 0.2

samples = pd.DataFrame(columns=feats_all)

for sample_id in (pbar := tqdm(data.index.values)):
    pbar.set_description(f"Sample {sample_id}")
    
    n_pos = 0
    n_neg = 0
    comp_present = []
    for comp in components:
        feats_w_trgt = components[comp]['feats'].index.values
        feats = feats_w_trgt[feats_w_trgt != feat_trgt]
        n_feats = len(feats)
        n_nans = data.loc[sample_id, feats].isna().sum()
        rho = components[comp]['metrics'].at['Test', 'pearson_corrcoef_unbiased'] * components[comp]['likelihood']
        if n_nans / n_feats < nan_part:
            comp_present.append(comp)
            data_sample = data.loc[[sample_id], feats_w_trgt]
            if n_nans != n_feats:
                data_bkcg = components[comp]['data'].loc[:, feats_w_trgt]
                data_imp = pd.concat([data_sample, data_bkcg], axis=0, ignore_index=True)
                imputer = KNNImputer(n_neighbors=5)
                data_sample.loc[sample_id, feats_w_trgt] = imputer.fit_transform(data_imp.loc[:, feats_w_trgt].values)[0, :]
            pred = components[comp]['model'].predict(data_sample)[f'{feat_trgt}_prediction'].values
            pred = components[comp]['corrector'].predict(pred)
            data_sample.at[sample_id, f"Предсказание {components[comp]['name']}"] = pred
            
            gt = data_sample.at[sample_id, feat_trgt]
            aa = pred - gt
            if aa > 0:
                n_pos += 1
            else:
                n_neg += 1
            
            data_sample.at[sample_id, f"Возрастная Акселерация {components[comp]['name']}"] = aa * rho
            
            samples.loc[sample_id, data_sample.columns] = data_sample.loc[sample_id, data_sample.columns]
            
    if len(comp_present) > 0:
        samples.at[sample_id, "Число моделей"] = len(comp_present)
        samples.at[sample_id, "Число моделей c отрицательной аккселерацией"] = n_neg
        samples.at[sample_id, "Число моделей c положительной аккселерацией"] = n_pos
        samples.at[sample_id, "Возрастная Акселерация"] = 0.0
        for comp in comp_present:
            samples.at[sample_id, f"Модель {components[comp]['name']}"] = True
            samples.at[sample_id, f"Возрастная Акселерация {components[comp]['name']}"] /= max(n_pos, n_neg)
            samples.at[sample_id, "Возрастная Акселерация"] += samples.at[sample_id, f"Возрастная Акселерация {components[comp]['name']}"]
        
samples.insert(len(samples.columns) - 1, "Число моделей", samples.pop("Число моделей"))
samples.insert(len(samples.columns) - 1, "Число моделей c отрицательной аккселерацией", samples.pop("Число моделей c отрицательной аккселерацией"))
samples.insert(len(samples.columns) - 1, "Число моделей c положительной аккселерацией", samples.pop("Число моделей c положительной аккселерацией"))
samples.insert(len(samples.columns) - 1, "Возрастная Акселерация", samples.pop("Возрастная Акселерация"))
samples['ATLAS Возраст'] = samples[feat_trgt] + samples["Возрастная Акселерация"]

samples.to_excel(f"{path}/AtlasAge/data___.xlsx")

## Load calculated data

samples = pd.read_excel(f"{path}/AtlasAge/data_full.xlsx", index_col=0)

## Prepare original data

samples_selected = pd.read_excel(f"{path}/AtlasAge/data_selected.xlsx", index_col=0)
data_raw = pd.read_excel(f"{path}/Выгрузка биовозраст 3.0_20241126.xlsx", index_col=0)
data_raw_selected = data_raw.loc[data_raw.index.intersection(samples_selected.index), :]
data_raw_selected.to_excel(f"{path}/AtlasAge/data_raw_selected.xlsx")

## Plot number of used models

models_count = samples['Число моделей'].value_counts().to_frame()
models_count.sort_index(inplace=True)
models_count['Количество моделей'] = models_count.index.astype(str)
models_count['Количество образцов'] = models_count['count']

exclude_colors = [components[comp]['color'] for comp in components] + ['black', 'white']
exclude_colors = [mcolors.hex2color(mcolors.CSS4_COLORS[color]) for color in exclude_colors]

colors = distinctipy.get_colors(models_count.shape[0], exclude_colors, rng=1337, pastel_factor=0.2)
colors_counts = {model_count: colors[model_count_id] for model_count_id, model_count in enumerate(models_count['Количество моделей'].values)}

barplot = sns.barplot(
    data=models_count,
    y='Количество моделей',
    x='Количество образцов',
    hue='Количество моделей',
    palette=colors_counts,
    edgecolor='black',
    legend=False,
)
for container in barplot.containers:
    barplot.bar_label(container, label_type='edge', fmt='%d', padding=2.5)
plt.savefig(f"{path}/AtlasAge/models_count.png", bbox_inches='tight', dpi=200)
plt.savefig(f"{path}/AtlasAge/models_count.pdf", bbox_inches='tight')
plt.close()


model_counts_dict = {
    f'Все образцы ({len(samples.index.values)})': samples.index.values,
    f"Больше 4 моделей ({len(samples.index[samples['Число моделей'] > 4].values)})": samples.index[samples['Число моделей'] > 4].values
}
model_counts_colors = {
    f'Все образцы ({len(samples.index.values)})': 'darkslategray',
    f"Больше 4 моделей ({len(samples.index[samples['Число моделей'] > 4].values)})": 'magenta',
}
for model_count_id, model_count in enumerate(models_count['Количество моделей'].values):
    model_count_index = samples.index[samples['Число моделей'] == int(model_count)].values
    model_counts_dict[f'Моделей = {model_count} (образцов = {len(model_count_index)})'] = model_count_index
    model_counts_colors[f'Моделей = {model_count} (образцов = {len(model_count_index)})'] = colors_counts[model_count]

for model_count, model_count_index in model_counts_dict.items():
    
    df_model_count = samples.loc[model_count_index, :]
    
    xy_min, xy_max = np.quantile(df_model_count[[feat_trgt, 'ATLAS Возраст']].values.flatten(), [0.01, 0.99])
    xy_ptp = xy_max - xy_min
    
    n_rows = 2
    n_cols = 2
    fig_height = 5
    fig_width = 7
    sns.set_theme(style='ticks')
    fig, axs = plt.subplots(n_rows, n_cols, figsize=(fig_width, fig_height), height_ratios=[2, 8],  width_ratios=[4, 2], gridspec_kw={'wspace':0.10, 'hspace': 0.05}, layout='constrained')

    mae = mean_absolute_error(df_model_count[feat_trgt].values, df_model_count['ATLAS Возраст'].values)
    rho, _ = stats.pearsonr(df_model_count[feat_trgt].values, df_model_count['ATLAS Возраст'].values)
    bias = np.mean(df_model_count['Возрастная Акселерация'].values)
    
    ds_table = pd.DataFrame(index=['Средняя абсолютная ошибка', 'Коэффициент корреляции Пирсона', 'Среднее смещение'], columns=[model_count])
    ds_table.at['Средняя абсолютная ошибка', model_count] = f"{mae:0.3f}"
    ds_table.at['Коэффициент корреляции Пирсона', model_count] = f"{rho:0.3f}"
    ds_table.at['Среднее смещение', model_count] = f"{bias:0.3f}"

    col_defs = [
        ColumnDefinition(
            name="index",
            title='Метрики',
            textprops={"ha": "left"},
            width=4.5,
        ),
        ColumnDefinition(
            name=model_count,
            textprops={"ha": "center"},
            width=2.0,
        ),
    ]
    table = Table(
        ds_table,
        column_definitions=col_defs,
        row_dividers=True,
        footer_divider=False,
        ax=axs[0, 0],
        textprops={"fontsize": 8},
        row_divider_kw={"linewidth": 1, "linestyle": (0, (1, 1))},
        col_label_divider_kw={"linewidth": 1, "linestyle": "-"},
        column_border_kw={"linewidth": 1, "linestyle": "-"},
    ).autoset_fontcolors(colnames=[model_count])

    kdeplot = sns.kdeplot(
        data=df_model_count,
        x=feat_trgt,
        y='ATLAS Возраст',
        fill=True,
        cbar=False,
        thresh=0.05,
        color=model_counts_colors[model_count],
        legend=False,
        ax=axs[1, 0]
    )
    bisect = sns.lineplot(
        x=[xy_min - 0.15 * xy_ptp, xy_max + 0.15 * xy_ptp],
        y=[xy_min - 0.15 * xy_ptp, xy_max + 0.15 * xy_ptp],
        linestyle='--',
        color='black',
        linewidth=1.0,
        ax=axs[1, 0]
    )
    regplot = sns.regplot(
        data=df_model_count,
        x=feat_trgt,
        y='ATLAS Возраст',
        color='k',
        scatter=False,
        truncate=False,
        ax=axs[1, 0]
    )
    axs[1, 0].set_xlim(xy_min - 0.15 * xy_ptp, xy_max + 0.15 * xy_ptp)
    axs[1, 0].set_ylim(xy_min - 0.15 * xy_ptp, xy_max + 0.15 * xy_ptp)
    axs[1, 0].set_ylabel("ATLAS возраст")
    axs[1, 0].set_xlabel(feat_trgt)
    
    axs[0, 1].axis('off')
    
    violin = sns.violinplot(
        data=df_model_count,
        x=[0] * df_model_count.shape[0],
        y='Возрастная Акселерация',
        color=model_counts_colors[model_count],
        density_norm='width',
        saturation=0.75,
        linewidth=1.0,
        ax=axs[1, 1],
        legend=False,
    )
    axs[1, 1].set_ylabel('Возрастная акселерация')
    axs[1, 1].set_xlabel('')
    axs[1, 1].set(xticklabels=[]) 
    axs[1, 1].set(xticks=[]) 
    fig.savefig(f"{path}/AtlasAge/models_count/{model_count}.png", bbox_inches='tight', dpi=200)
    fig.savefig(f"{path}/AtlasAge/models_count/{model_count}.pdf", bbox_inches='tight')
    plt.close(fig)
    


components_n_samples = pd.DataFrame(index=list(components.keys()))
for comp in components:
    components_n_samples.at[comp, 'Count'] = components[comp]['data'].shape[0]
components_n_samples.sort_values(by='Count', inplace=True)

dict_upset_lists = {}
for comp in components_n_samples.index.values:
    dict_upset_lists[f"                 {components[comp]['name']}"] = samples.index[samples[f"Модель {components[comp]['name']}"] == True].values
upset_all = list(set().union(*list(dict_upset_lists.values())))
df_upset = pd.DataFrame(index=upset_all)
for k, v in dict_upset_lists.items():
    df_upset[k] = df_upset.index.isin(v)
df_upset = df_upset.set_index(list(dict_upset_lists.keys()))
tmp = plt.figure(figsize=(50, 9))
upset_fig = upsetplot.UpSet(
    df_upset,
    sort_categories_by='input',
    subset_size='count',
    show_counts=True,
    min_degree=0,
    element_size=None,
    totals_plot_elements=3,
    include_empty_subsets=False
)
upset_fig.plot(tmp)
plt.savefig(f"{path}/AtlasAge/models_upset.png", bbox_inches='tight')
plt.savefig(f"{path}/AtlasAge/models_upset.pdf", bbox_inches='tight')
plt.close()

# Local explainability

trgt_id = 681241 # 14304 144

local_exlp = {}

n_pos = samples.at[trgt_id, "Число моделей c положительной аккселерацией"]
n_neg = samples.at[trgt_id, "Число моделей c отрицательной аккселерацией"]

for comp in components:
    
    if samples.at[trgt_id, f"Модель {components[comp]['name']}"] == True:
        feats_comp = components[comp]['feats']
        feats_w_trgt = components[comp]['feats'].index.values
        feats = feats_w_trgt[feats_w_trgt != feat_trgt]
        feats_corr = components[comp]['feats_corr']
        metrics = components[comp]['metrics']
        data_shap = components[comp]['data_shap']
        
        def predict_func(X):
            X_df = pd.DataFrame(data=X, columns=feats)
            y = components[comp]['model'].predict(X_df)[f'{feat_trgt}_prediction'].values
            y = components[comp]['corrector'].predict(y)
            return y
        
        mae = metrics.at['Test', 'mean_absolute_error_unbiased']
        rho = metrics.at['Test', 'pearson_corrcoef_unbiased'] * components[comp]['likelihood']
        
        color = components[comp]['color']
        bkg_count = components[comp]['bkg_count']
        
        trgt_age = samples.at[trgt_id, feat_trgt]
        trgt_pred_raw = samples.at[trgt_id, f"Предсказание {components[comp]['name']}"]
        
        data_closest = data_shap.loc[data_shap['Error Unbiased'].abs() < mae * rho, :]
        data_closest = data_closest.iloc[(data_closest['Prediction Unbiased'] - trgt_age).abs().argsort()[:bkg_count]]
        # data_closest = data_shap.iloc[(data_shap['Prediction Unbiased'] - trgt_age).abs().argsort()[:bkg_count]]

        explainer = shap.SamplingExplainer(predict_func, data_closest.loc[:, feats].values)
        shap_values = explainer.shap_values(samples.loc[[trgt_id], feats].values)[0]
        shap_values = shap_values * (trgt_pred_raw - trgt_age) / (trgt_pred_raw - explainer.expected_value)
        print(f"{sum(shap_values) - (trgt_pred_raw - trgt_age)}")
        shap_values = shap_values * rho / max(n_pos, n_neg)
        
        
        # SHAP values correction
        shap_corr_thld = 5.0
        shap_corr_to = 1.0
        shap_mean_abs = np.mean(np.abs(shap_values))
        if shap_mean_abs > shap_corr_thld and abs(sum(shap_values)) < shap_corr_thld:
            print('SHAP values correction')
            shap_pos_ids = np.squeeze(np.argwhere(shap_values >= 0))
            shap_neg_ids = np.squeeze(np.argwhere(shap_values < 0))
            
            shap_pos_sum_abs = np.sum(np.abs(shap_values[shap_pos_ids]))
            shap_neg_sum_abs = np.sum(np.abs(shap_values[shap_neg_ids]))
            
            shap_sum_abs_from = np.sum(np.abs(shap_values))
            shap_sum_abs_to = shap_corr_to * len(shap_values)
            
            shap_corr_diff = shap_sum_abs_from - shap_sum_abs_to
            
            for pos_id in shap_pos_ids:
                curr_part = abs(shap_values[pos_id]) / shap_pos_sum_abs
                shap_values[pos_id] -= curr_part * shap_corr_diff * 0.5
            for neg_id in shap_neg_ids:
                curr_part = abs(shap_values[neg_id]) / shap_neg_sum_abs
                shap_values[neg_id] += curr_part * shap_corr_diff * 0.5
        
        
        df_comp = pd.DataFrame(index=feats, columns=['SHAP', 'Values', 'Correlation', 'Percentile', 'Class', 'Consistent', 'Show'])
        df_comp['SHAP'] = shap_values
        df_comp.sort_values(by='SHAP', key=abs, inplace=True)
        df_comp.loc[df_comp.index.values, 'Values'] = samples.loc[trgt_id, df_comp.index.values].values
        df_comp.loc[df_comp.index.values, 'Correlation'] = feats_corr.loc[df_comp.index.values, 'Correlation']
        for f in df_comp.index.values:
            df_comp.at[f, 'Percentile'] = scipy.stats.percentileofscore(data_closest.loc[:, f].values, samples.at[trgt_id, f])
            df_comp.at[f, 'Class'] = int(df_comp.at[f, 'Percentile'] // 33.333334)
            if (
                ((df_comp.at[f, 'Correlation'] > 0) & (df_comp.at[f, 'SHAP'] > 0) & (df_comp.at[f, 'Class'] == 2)) or \
                ((df_comp.at[f, 'Correlation'] > 0) & (df_comp.at[f, 'SHAP'] < 0) & (df_comp.at[f, 'Class'] == 0)) or \
                ((df_comp.at[f, 'Correlation'] < 0) & (df_comp.at[f, 'SHAP'] > 0) & (df_comp.at[f, 'Class'] == 0)) or \
                ((df_comp.at[f, 'Correlation'] < 0) & (df_comp.at[f, 'SHAP'] < 0) & (df_comp.at[f, 'Class'] == 2))
                ):
                df_comp.at[f, 'Consistent'] = True
            else:
                df_comp.at[f, 'Consistent'] = False
            if (
                (df_comp.at[f, 'Consistent'] == True) or \
                ((df_comp.at[f, 'Consistent'] == False) & (df_comp.at[f, 'Class'] == 1))
                ):
                df_comp.at[f, 'Show'] = True
            else:
                df_comp.at[f, 'Show'] = False
        df_comp['Class'] = df_comp['Class'].replace({0: 'Понижен', 1: 'Средний', 2: 'Повышен'})
        
        local_exlp[comp] = {
            'df': df_comp,
            'age_acceleration': samples.at[trgt_id, f"Возрастная Акселерация {components[comp]['name']}"],
        }

df_comps = pd.DataFrame(index=list(local_exlp.keys()), columns=['age_acceleration'])    
for comp in local_exlp:
    df_comps.at[comp, 'age_acceleration'] = local_exlp[comp]['age_acceleration']
df_comps['abs_age_acceleration'] = df_comps['age_acceleration'].abs()
df_comps.sort_values(["abs_age_acceleration"], ascending=[True], inplace=True)
df_comps['cumsum'] = df_comps['age_acceleration'].cumsum()
comps_sorted = df_comps.index.values

trgt_aa = df_comps['age_acceleration'].sum()
trgt_age = samples.at[trgt_id, feat_trgt]

summary_hover_text = ["Хронологический возраст"] + [f"{components[comp]['name']}" for comp in comps_sorted] + ["Биологический возраст"]
summary_measure = ['absolute'] + ['relative'] * len(comps_sorted) + ['absolute']
summary_ys = [-0.5] + [x + 0.5 for x in range(len(comps_sorted))] + [len(comps_sorted) + 0.5]
aas = df_comps['age_acceleration'].tolist()
summary_xs = [trgt_age] + aas + [trgt_age + trgt_aa]
summary_text = [f"{trgt_age:0.2f}"] + [f"+{aa:0.2f}" if aa > 0 else f"{aa:0.2f}" for aa in aas] + [f"{trgt_age+trgt_aa:0.2f}"]

fig = go.Figure()
fig.add_trace(
    go.Waterfall(
        hovertext=summary_hover_text,
        orientation="h",
        measure=summary_measure,
        y=summary_ys,
        x=summary_xs,
        base=0,
        text=summary_text,
        textposition = "auto",
        decreasing = {"marker":{"color": "deepskyblue", "line": {"color": "black", "width": 1}}},
        increasing = {"marker":{"color": "crimson", "line": {"color": "black", "width": 1}}},
        totals= {"marker":{"color": "dimgray", "line": {"color": "black", "width": 1}}},
        connector={
            "mode": "between",
            "line": {"width": 1, "color": "black", "dash": "dot"},
        },
    ),
)
fig.update_yaxes(
    automargin=True,
    tickmode="array",
    tickvals=summary_ys,
    ticktext=summary_hover_text,
    tickfont=dict(size=18),
)
fig.update_xaxes(
    automargin=True,
    title='Возраст',
    titlefont=dict(size=20),
    range=[
        trgt_age + df_comps['cumsum'].min() * 1.2 - 2,
        trgt_age + df_comps['cumsum'].max() * 1.2 + 2
    ],
)
fig.update_layout(
    title=f"Возрастная акселерация для {trgt_id}",
    titlefont=dict(size=25),
    template="none",
    width=800,
    height=len(comps_sorted)*30 + 200,
    margin=go.layout.Margin(l=120, r=100, b=50, t=50, pad=0),
)
fig.show()

for comp in comps_sorted[::-1]:

    df_comp = local_exlp[comp]['df']
    aa_comp = local_exlp[comp]['age_acceleration']

    df_comp.sort_values(['SHAP'], key=abs, inplace=True)
    df_comp['SHAP cumsum'] = df_comp['SHAP'].cumsum()

    n_show = 5
    
    fig = make_subplots(rows=1, cols=2, shared_yaxes=True, shared_xaxes=False, column_widths=[2.5, 0.5], horizontal_spacing=0.15, subplot_titles=['', 'Корреляция с возрастом'])

    if df_comp.shape[0] <= n_show:

        fig.add_trace(
            go.Waterfall(
                hovertext=["Итоговая акселерация"],
                orientation="h",
                measure=['relative'],
                y=[df_comp.shape[0] + 1],
                x=[aa_comp],
                base=0,
                text=[f"+{aa_comp:0.2f}" if aa_comp > 0 else f"{aa_comp:0.2f}"],
                textposition = "auto",
                decreasing = {"marker":{"color": "deepskyblue", "line": {"color": "black", "width": 1}}},
                increasing = {"marker":{"color": "crimson", "line": {"color": "black", "width": 1}}},
                totals= {"marker":{"color": "dimgray", "line": {"color": "black", "width": 1}}},
                connector={
                    "mode": "between",
                    "line": {"width": 1, "color": "black", "dash": "dot"},
                },
            ),
            row=1,
            col=1,
        )
        fig.add_trace(
            go.Waterfall(
                hovertext=df_comp.index.values,
                orientation="h",
                measure=["relative"] * len(feats),
                y=list(range(df_comp.shape[0])),
                x=df_comp['SHAP'].values,
                base=0,
                text=[f"+{x:0.2f}" if x > 0 else f"{x:0.2f}" for x in df_comp['SHAP'].values],
                textposition = "auto",
                decreasing = {"marker":{"color": "lightblue", "line": {"color": "black", "width": 1}}},
                increasing = {"marker":{"color": "lightcoral", "line": {"color": "black", "width": 1}}},
                connector={
                    "mode": "between",
                    "line": {"width": 1, "color": "black", "dash": "solid"},
                },
            ),
            row=1,
            col=1,
        )
        fig.update_traces(row=1, col=1, showlegend=False)
        fig.update_yaxes(
            row=1,
            col=1,
            automargin=True,
            tickmode="array",
            tickvals=list(range(df_comp.shape[0])) + [df_comp.shape[0] + 1],
            ticktext=[f"{x} = {df_comp.at[x, 'Values']:0.2f} ({df_comp.at[x, 'Class']})" for x in df_comp.index] + ['Итоговая акселерация'],
            tickfont=dict(size=18),
        )
        fig.update_xaxes(
            row=1,
            col=1,
            automargin=True,
            title='Возрастная акселерация',
            titlefont=dict(size=20),
        )

        fig.add_trace(
            go.Heatmap(
                x=['Возраст'],
                y=list(range(df_comp.shape[0])),
                z=[[x] for x in df_comp['Correlation'].values],
                text=[[f"{x:0.2f}"] for x in df_comp['Correlation'].values],
                texttemplate="%{text}",
                zmin=-1,
                zmax=1,
                colorbar_y=df_comp.shape[0] / (df_comp.shape[0] + 2) * 0.5,
            ),
            row=1,
            col=2,
        )

        fig.update_layout(
            title=f"{components[comp]['name']}",
            titlefont=dict(size=25),
            template="none",
            width=1800,
            height=df_comp.shape[0]*40 + 100,
            margin=go.layout.Margin(l=120, r=100, b=50, t=50, pad=0),
        )
    else:
        
        feats_show = df_comp.index[df_comp['Show'] == True].tolist()
        if len(feats_show) >= n_show:
            feats_show = feats_show[-n_show:]
        
        df_show = df_comp.loc[feats_show, :]
        df_not_show = df_comp.loc[list(set(df_comp.index.values) - set(feats_show)), :]
        df_show.sort_values(['SHAP'], key=abs, inplace=True)
        df_not_show.sort_values(['SHAP'], key=abs, inplace=True)
        shap_not_show_sum = df_not_show['SHAP'].sum()

        fig.add_trace(
            go.Waterfall(
                hovertext=["Итоговая акселерация"],
                orientation="h",
                measure=['relative'],
                y=[df_show.shape[0] + 1],
                x=[aa_comp],
                base=0,
                text=[f"+{aa_comp:0.2f}" if aa_comp > 0 else f"{aa_comp:0.2f}"],
                textposition = "auto",
                decreasing = {"marker":{"color": "deepskyblue", "line": {"color": "black", "width": 1}}},
                increasing = {"marker":{"color": "crimson", "line": {"color": "black", "width": 1}}},
                totals= {"marker":{"color": "dimgray", "line": {"color": "black", "width": 1}}},
                connector={
                    "mode": "between",
                    "line": {"width": 1, "color": "black", "dash": "dot"},
                },
            ),
            row=1,
            col=1,
        )
        fig.add_trace(
            go.Waterfall(
                hovertext=df_show.index.values,
                orientation="h",
                measure=["relative"] * len(feats),
                y=[-1] + list(range(df_show.shape[0])),
                x=[shap_not_show_sum] + df_show['SHAP'].tolist(),
                base=0,
                text=[f"+{shap_not_show_sum:0.2f}" if shap_not_show_sum > 0 else f"{shap_not_show_sum:0.2f}"] + [f"+{x:0.2f}" if x > 0 else f"{x:0.2f}" for x in df_show['SHAP'].values],
                textposition = "auto",
                decreasing = {"marker":{"color": "lightblue", "line": {"color": "black", "width": 1}}},
                increasing = {"marker":{"color": "lightcoral", "line": {"color": "black", "width": 1}}},
                connector={
                    "mode": "between",
                    "line": {"width": 1, "color": "black", "dash": "solid"},
                },
            ),
            row=1,
            col=1,
        )
        fig.update_traces(row=1, col=1, showlegend=False)
        fig.update_yaxes(
            row=1,
            col=1,
            automargin=True,
            tickmode="array",
            tickvals=[-1] + list(range(df_show.shape[0])) + [df_show.shape[0] + 1],
            ticktext=['Остальные признаки'] + [f"{x} = {df_show.at[x, 'Values']:0.2f} ({df_show.at[x, 'Class']})" for x in df_show.index] + ['Итоговая акселерация'],
            tickfont=dict(size=18),
        )
        fig.update_xaxes(
            row=1,
            col=1,
            automargin=True,
            title='Возрастная акселерация',
            titlefont=dict(size=20),
        )

        fig.add_trace(
            go.Heatmap(
                x=['Возраст'],
                y=list(range(df_show.shape[0])),
                z=[[x] for x in df_show['Correlation'].values],
                text=[[f"{x:0.2f}"] for x in df_show['Correlation'].values],
                texttemplate="%{text}",
                zmin=-1,
                zmax=1,
                colorbar_y=(df_show.shape[0] + 1) / (df_show.shape[0] + 3) * 0.5,
            ),
            row=1,
            col=2,
        )

        fig.update_layout(
            title=f"{components[comp]['name']}",
            titlefont=dict(size=25),
            template="none",
            width=1800,
            height=(df_show.shape[0] + 1) * 40 + 100,
            margin=go.layout.Margin(l=120, r=100, b=50, t=50, pad=0),
        )
    
    fig.update_xaxes(
        row=1,
        col=2,
        tickvals=[],
        automargin=True,
        showgrid=False,
        showline=False,
        zeroline=False,
        showticklabels=False,
    )
    fig.update_yaxes(
        row=1,
        col=2,
        tickvals=[],
        automargin=True,
        showgrid=False,
        showline=False,
        zeroline=False,
        showticklabels=False,
    )
    fig.show()

# Gradio theme

import gradio as gr

primary_hue = mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 1.00), keep_alpha=True)

color_primary = gr.themes.Color(
    c50=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.05), keep_alpha=True),
    c100=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.1), keep_alpha=True),
    c200=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.2), keep_alpha=True),
    c300=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.3), keep_alpha=True),
    c400=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.4), keep_alpha=True),
    c500=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.5), keep_alpha=True),
    c600=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.6), keep_alpha=True),
    c700=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.7), keep_alpha=True),
    c800=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.8), keep_alpha=True),
    c900=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.9), keep_alpha=True),
    c950=mcolors.rgb2hex(make_rgb_transparent(mcolors.to_rgb("#005450"), (1, 1, 1), 0.95), keep_alpha=True),
    name='PANTONE 7722 C'
)

import gradio as gr

gr.themes.colors.gray.c100

# Generate docx report

from docx.oxml.ns import qn

path = f"E:/YandexDisk/Work/bbd/atlas/AtlasAge"

doc = Document()

for s in doc.styles:
    print(s.name)

doc.styles['Title'].font.size = Pt(24)
doc.styles['Title'].font.name = 'Euclid Square'
doc.styles['Title'].font.bold = True
doc.styles['Title'].font.color.rgb = RGBColor(0, 84, 80)
doc.styles['Title'].paragraph_format._element.xpath('.//w:pBdr')[0].getparent().remove(doc.styles['Title'].paragraph_format._element.xpath('.//w:pBdr')[0])
doc.styles['Title'].element.rPr.rFonts.set(qn("w:asciiTheme"), "Euclid Square")
doc.styles['Title'].element.rPr.rFonts.set(qn("w:hAnsiTheme"), "Euclid Square")
doc.styles['Title'].element.rPr.rFonts.set(qn('w:eastAsia'), 'Euclid Square')

doc.styles['Heading 1'].font.size = Pt(16)
doc.styles['Heading 1'].font.name = 'Euclid Square'
doc.styles['Heading 1'].font.bold = True
doc.styles['Heading 1'].font.color.rgb = RGBColor(0, 84, 80)
doc.styles['Heading 1'].element.rPr.rFonts.set(qn("w:asciiTheme"), "Euclid Square")
doc.styles['Heading 1'].element.rPr.rFonts.set(qn("w:hAnsiTheme"), "Euclid Square")
doc.styles['Heading 1'].element.rPr.rFonts.set(qn('w:eastAsia'), 'Euclid Square')

doc.styles['Normal'].font.size = Pt(13)
doc.styles['Normal'].font.name = 'Euclid Square'
doc.styles['Normal'].font.color.rgb = RGBColor(0, 0, 0)

sections = doc.sections
for section in sections:
    section.page_width = Mm(210)
    section.page_height = Mm(297)
    section.top_margin = Mm(10)
    section.bottom_margin = Mm(10)
    section.left_margin = Mm(15)
    section.right_margin = Mm(15)

title = doc.add_paragraph('Биовозраст АТЛАС', style='Title')
title.alignment = WD_ALIGN_PARAGRAPH.CENTER
title.bold = True

paragraph = doc.add_paragraph(style='Normal')
paragraph.alignment = WD_ALIGN_PARAGRAPH.CENTER
run = paragraph.add_run(f"Пациент: ФИО")
run.font.size=Pt(16)
run.font.bold = True

table = doc.add_table(rows=1, cols=3)
delete_paragraph(table.cell(0, 0).paragraphs[0])
paragraph = table.cell(0, 0).add_paragraph(style='Normal')
paragraph.alignment = WD_ALIGN_PARAGRAPH.CENTER
run = paragraph.add_run(f'Возраст: 33')
run.font.bold = True
delete_paragraph(table.cell(0, 1).paragraphs[0])
paragraph = table.cell(0, 1).add_paragraph(style='Normal')
paragraph.alignment = WD_ALIGN_PARAGRAPH.CENTER
run = paragraph.add_run(f'Биовозраст АТЛАС: 33')
run.font.color.rgb = RGBColor(0, 84, 80)
run.font.bold = True
delete_paragraph(table.cell(0, 2).paragraphs[0])
paragraph = table.cell(0, 2).add_paragraph(style='Normal')
paragraph.alignment = WD_ALIGN_PARAGRAPH.CENTER
run = paragraph.add_run(f'Акселерация: 33')
run.font.color.rgb = RGBColor(255, 0, 0)
run.font.bold = True

heading_1 = doc.add_paragraph('Общий результат по всем группам анализов:', style='Heading 1')
heading_1.alignment = WD_ALIGN_PARAGRAPH.CENTER
heading_1.bold = True

paragraph = doc.add_paragraph(style='Normal')
paragraph.alignment = WD_ALIGN_PARAGRAPH.CENTER
paragraph.add_run().add_picture(f"{path}/docx/AgeAcceleration_global.png", width=Mm(150))

heading_1 = doc.add_paragraph('Результаты по отдельным группам анализов:', style='Heading 1')
heading_1.alignment = WD_ALIGN_PARAGRAPH.CENTER
heading_1.bold = True

for comp_id in range(8):
    paragraph = doc.add_paragraph(style='Normal')
    paragraph.alignment = WD_ALIGN_PARAGRAPH.CENTER
    paragraph.add_run().add_picture(f"{path}/docx/AgeAcceleration_{comp_id}.png", width=Mm(190))

paragraph = doc.add_paragraph(style='Normal')

doc.save(f"{path}/docx/Отчет.docx")
